class and GPT
I tried to create a class with x any coordinates in matplotlib. But I failed really badly.  I tried to look it up but did not find anything.
I used chat GPT to create the class for me which it did. Now i will use it to create the stars.