Final Simulation
Thanks to the help of rebound and the amazing Chat GPT I have been able to create two simulation. One for animating the orbits and another for drawing the orbits over time. This should be enough to star working to produce results. I hope to be done with the results part tomorrow :100: